Delete grades by id

+++ src/main/java/com/luv2code/springmvc/repository/ScienceGradeDao.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ScienceGradeDao extends CrudRepository<ScienceGrade, Long> {
+public interface ScienceGradeDao extends CrudRepository<ScienceGrade, Integer> {
 
     Iterable<ScienceGrade> findGradeByStudentId(int i);
 }
